requirements.txt
python-telegram-bot>=21.0.1
git+https://github.com/openai/whisper.git
langdetect==1.0.9
deep-translator==1.11.4
gTTS==2.4.0
torch>=2.0.0 --extra-index-url https://download.pytorch.org/whl/cu118
numpy>=1.20.0
python-dotenv>=1.0.0 